+++ CHANGES.rst
@@ -19,6 +19,9 @@ Unreleased
 - Send Signals for cache hits and misses. :pr:`#237` and :pr:`667`
 - Include ``key_prefix`` when building ``@cached(query_string=True)`` cache
   keys. :issue:`302`
+- Add ``Cache.delete_cached()`` and extend ``make_cache_key()`` with ``path`` and ``query_args``
+  arguments to make deleting views decorated with ``cached(query_string=True)`` possible.
+  :issue:`243`
 - Use ``hashlib.sha256`` instead of ``hashlib.md5`` for hashing the cache keys.
   This changes the generated keys, so entries cached by an earlier version become
   obsolete. If you wish to still use hashlib.md5 set the config